+++ IbraheemKarim.MathGame/IbraheemKarim.MathGame.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.9.34622.214
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IbraheemKarim.MathGame", "IbraheemKarim.MathGame\IbraheemKarim.MathGame.csproj", "{6ECB55A8-1CB2-468F-941F-68EE558DC7B3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Original", "Original\Original.csproj", "{81BC47E1-28EF-449E-B95C-76B43C49FEDB}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{6ECB55A8-1CB2-468F-941F-68EE558DC7B3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6ECB55A8-1CB2-468F-941F-68EE558DC7B3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6ECB55A8-1CB2-468F-941F-68EE558DC7B3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{81BC47E1-28EF-449E-B95C-76B43C49FEDB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{81BC47E1-28EF-449E-B95C-76B43C49FEDB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{81BC47E1-28EF-449E-B95C-76B43C49FEDB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{81BC47E1-28EF-449E-B95C-76B43C49FEDB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
