+++ IbraheemKarim.MathGame/IbraheemKarim.MathGame/GameEngine.cs
@@ -7,7 +7,6 @@
             GameType gameType = GetDesiredGameType();
             ProcessGameTypeChoice(gameType);
         }
-
         private GameType GetDesiredGameType()
         {
             bool firstIteration = true;
@@ -36,7 +35,6 @@ Choose what kind of game would you like to play:
             } while (true);
 
         }
-
         private void ProcessGameTypeChoice(GameType gameType)
         {
             switch (gameType)
@@ -51,12 +49,11 @@ Choose what kind of game would you like to play:
                     this.CustomMathGame("Multiplication game", '*', (first, second) => first * second, GameType.Multiplication);
                     break;
                 case GameType.Division:
-                    //this.DivisionGame("Division game");
+                    this.DivisionGame("Division game");
                     break;
             }
 
         }
-
         private void CustomMathGame(string gameTitle, char operationSymbol, Func<int, int, int> operationFunction, GameType gameType)
         {
             var random = new Random();
@@ -97,7 +94,43 @@ Choose what kind of game would you like to play:
 
             GamesHistory.AddToHistory(score, gameType);
         }
+        private void DivisionGame(string message)
+        {
+            var score = 0;
+
+            for (int i = 0; i < 5; i++)
+            {
+                Console.Clear();
+                Console.WriteLine(message);
+
+                var divisionNumbers = Helpers.GetDivisionNumbers();
+                var firstNumber = divisionNumbers[0];
+                var secondNumber = divisionNumbers[1];
+
+                Console.WriteLine($"{firstNumber} / {secondNumber}");
+
+                var result = Console.ReadLine();
+                result = Helpers.ValidateResult(result);
+
+                if (int.Parse(result) == firstNumber / secondNumber)
+                {
+                    Console.WriteLine("Your answer was correct! Press any key to continue");
+                    score++;
+                }
+                else
+                {
+                    Console.WriteLine("Your answer was incorrect. Press any key to continue");
+                }
+
+                Console.ReadKey();
+            }
 
 
+            Console.Clear();
+            Console.WriteLine($"Game over. Your final score is {score}/5. Press any key to go back to the main menu.");
+            Console.ReadKey();
+
+            GamesHistory.AddToHistory(score, GameType.Division);
+        }
     }
 }
