+++ IbraheemKarim.MathGame/IbraheemKarim.MathGame/GameRecord.cs
@@ -1,19 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace IbraheemKarim.MathGame
+﻿namespace IbraheemKarim.MathGame
 {
     internal class GameRecord
     {
         public DateTime Date { get; set; }
-
         public int Score { get; set; }
-
         public GameType Type { get; set; }
-
 
         public override string ToString()
         {
