+++ IbraheemKarim.MathGame/IbraheemKarim.MathGame/Helpers.cs
@@ -1,5 +1,4 @@
-﻿
-namespace IbraheemKarim.MathGame
+﻿namespace IbraheemKarim.MathGame
 {
     internal static class Helpers
     {
@@ -15,13 +14,32 @@ namespace IbraheemKarim.MathGame
 
             return name;
         }
-        internal static string ValidateResult(string? result)
+        public static string ValidateResult(string? result)
         {
             while (string.IsNullOrEmpty(result) || !Int32.TryParse(result, out _))
             {
                 Console.WriteLine("Invalid input. Try again.");
                 result = Console.ReadLine();
             }
+
+            return result;
+        }
+        public static int[] GetDivisionNumbers()
+        {
+            var random = new Random();
+            var firstNumber = random.Next(1, 99);
+            var secondNumber = random.Next(1, 99);
+
+            var result = new int[2];
+
+            while (firstNumber % secondNumber != 0)
+            {
+                firstNumber = random.Next(1, 99);
+                secondNumber = random.Next(1, 99);
+            }
+
+            result[0] = firstNumber;
+            result[1] = secondNumber;
 
             return result;
         }
