+++ IbraheemKarim.MathGame/IbraheemKarim.MathGame/Program.cs
@@ -1,6 +1,5 @@
 ﻿using IbraheemKarim.MathGame;
 
-
 string username = Helpers.GetUserName();
 
 var menu = new Menu(new GameEngine());
