--- IbraheemKarim.MathGame/Original/Menu.cs
@@ -1,59 +1,0 @@
-﻿namespace MyFirstProgram;
-
-internal class Menu
-{
-    GameEngine gamesClass = new();
-
-    internal void ShowMenu(string name, DateTime date)
-    {
-        Console.Clear();
-        Console.WriteLine($"Hello {name}. It's {date}. This is your math's game. That's great that you're working on improving yourself");
-        Console.WriteLine("Press any key to show menu");
-        Console.ReadLine();
-        Console.WriteLine("\n");
-
-        var isGameOn = true;
-
-        do
-        {
-            Console.Clear();
-            Console.WriteLine(@$"
-What game would you like to play today? Choose from the options below:
-V - View Previous Games
-A - Addition
-S - Subtraction
-M - Multiplication
-D - Division
-Q - Quit the program");
-            Console.WriteLine("---------------------------------------------");
-
-            var gameSelected = Console.ReadLine();
-
-            switch (gameSelected.Trim().ToLower())
-            {
-                case "v":
-                    Helpers.PrintGames();
-                    break;
-                case "a":
-                    gamesClass.AdditionGame("Addition game");
-                    break;
-                case "s":
-                    gamesClass.SubtractionGame("Subtraction game");
-                    break;
-                case "m":
-                    gamesClass.MultiplicationGame("Multiplication game");
-                    break;
-                case "d":
-                    gamesClass.DivisionGame("Division game");
-                    break;
-                case "q":
-                    Console.WriteLine("Goodbye");
-                    isGameOn = false;
-                    break;
-                default:
-                    Console.WriteLine("Invalid Input");
-                    break;
-            }
-        } while (isGameOn);
-    }
-}
