--- IbraheemKarim.MathGame/Original/Models/Game.cs
@@ -1,25 +1,0 @@
-﻿namespace Academy.Console.MathGame.Models;
-
-internal class Game
-{
-    //private int _score;
-
-    //public int Score 
-    //{ 
-    //    get { return _score; } 
-    //    set { _score = value; }
-    //}
-    public DateTime Date { get; set; }
-
-    public int Score { get; set; }
-
-    public GameType Type { get; set; }
-}
-
-internal enum GameType
-{
-    Addition,
-    Subtraction,
-    Multiplication,
-    Division
-}
