--- IbraheemKarim.MathGame/Original/Program.cs
@@ -1,11 +1,0 @@
-﻿using MyFirstProgram;
-
-
-var date = DateTime.UtcNow;
-
-string name = Helpers.GetName();
-
-var menu = new Menu();
-
-
-menu.ShowMenu(name, date);
